Remove Extent::for_each*
These were redundant with Extent::iter*, and there was no perf difference.

The reason we had these in the first place was that we knew itertools::iproduct
was slower than a nested for loop, but the flat_map-based iterator seems to be
much faster, so we'll just always use that.